+++ CHANGELOG.md
@@ -2,15 +2,29 @@
 
 All notable changes to this project will be documented in this file.
 
-## [unreleased]
+## [0.2.0] - 14-12-2025
 
 ### Added
 - Authenticated navigation bar
 - `authService` now makes an API call to sync user credentials upon login
+- Group features in the dashboard page
+  - Group creation
+  - Group setting modification features (change name/desc)
+  - Group member features (View & kick members, view & revoke invitations, send invitation to group)
+  - Group deletion feature
+  - Group leave feature
+- `useApiErrorHandler` hook to handle API errors
+
+### Removed
+- Removed `Zustand` as a dependency
 
 ### Fixed
 - Resolved an issue with `AuthContext` that was causing logged in users to be stuck at "Loading..." page
 - Resolved an issue with `AuthContext` that was repeatedly syncing user information with the server
+- Downgraded `Zod` version due to some potential bug with `react-hook-form`
+
+### Infrastructure
+- Set up Supabase trigger to check whitelist table during first time login, rejects if email is not whitelisted
 
 ## [0.1.1] - 12-11-2025